requirements.txt
Django>=4.2,<5.0
djangorestframework
django-environ
psycopg2-binary
drf-spectacular
celery
redis
gunicorn
django-cors-headers
whitenoise
djangorestframework-simplejwt
